plugin-EnergyCampus: decouple rest lib from ennovatis for later usage in web plugin
