+++ src/OpenCOVER/plugins/hlrs/Energy/CMakeLists.txt
@@ -17,7 +17,7 @@ SET(SOURCES
 
 add_subdirectory(utils)
 add_subdirectory(ennovatis)
-cover_add_plugin(EnergyCampus $<TARGET_OBJECTS:Ennovatis> $<TARGET_OBJECTS:EnergyUtils> ${HEADERS} ${SOURCES})
+cover_add_plugin(EnergyCampus $<TARGET_OBJECTS:EnergyUtils> $<TARGET_OBJECTS:Ennovatis> ${HEADERS} ${SOURCES})
 target_include_directories(EnergyCampus PRIVATE
 	"${CMAKE_CURRENT_BINARY_DIR}/ennovatis/include"
     PRIVATE "${CMAKE_CURRENT_BINARY_DIR}/utils/include")
