+++ src/OpenCOVER/plugins/hlrs/Energy/EnnovatisDevice.h
@@ -16,7 +16,7 @@
 class EnnovatisDevice {
 public:
     EnnovatisDevice(const ennovatis::Building &building): EnnovatisDevice(building, nullptr){};
-    EnnovatisDevice(const ennovatis::Building &building, std::shared_ptr<ennovatis::RESTRequest> req);
+    EnnovatisDevice(const ennovatis::Building &building, std::shared_ptr<ennovatis::rest_request> req);
     ~EnnovatisDevice();
     [[nodiscard]] bool getStatus() { return m_InfoVisible; }
     [[nodiscard]] osg::ref_ptr<osg::Group> getDeviceGroup() { return m_deviceGroup; }
@@ -43,7 +43,7 @@ private:
 
     float m_rad;
     bool m_InfoVisible = false;
-    std::shared_ptr<ennovatis::RESTRequest> m_request;
+    std::shared_ptr<ennovatis::rest_request> m_request;
     std::shared_ptr<ennovatis::ChannelGroup> m_channelGroup;
     BuildingInfo m_buildingInfo;
 };
