+++ src/OpenCOVER/plugins/hlrs/Energy/utils/CMakeLists.txt
@@ -1,4 +1,5 @@
 USING(CURL)
 file(GLOB_RECURSE SOURCES "*.cpp" "*.h")
 add_library(EnergyUtils OBJECT ${SOURCES})
-set_target_properties(EnergyUtils PROPERTIES LINKER_LANGUAGE CXX)
+target_link_libraries(EnergyUtils ${CURL_LIBRARIES})
+set_target_properties(EnergyUtils PROPERTIES LINKER_LANGUAGE CXX POSITION_INDEPENDENT_CODE ON)
